% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.nice.R
\name{round.nice}
\alias{round.nice}
\title{Nice Round}
\usage{
\method{round}{nice}(x, nice = c(1:10))
}
\arguments{
\item{x}{Most numbers. I'm not sure how robust this is}
}
\description{
Rounds numbers to a "pretty" number
}
\examples{
round.nice()
}

