% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.cibox.R
\name{sas.cibox}
\alias{sas.cibox}
\title{Confidence Interval Boxplot}
\usage{
sas.cibox(dataset, alpha = 0.05, axisTF = TRUE)
}
\arguments{
\item{dataset}{The data to plot}

\item{alpha}{The confidence level. Defaults to .05}

\item{axisTF}{Determines if the plot has axes or not. Used to simplify sas.tplots()}
}
\description{
This creates a boxplot with the confidence interval highlighted
}
\examples{
sas.cibox()
}

