% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.densityplot.R
\name{sas.densityplot}
\alias{sas.densityplot}
\title{Density Plots}
\usage{
sas.densityplot(dataset)
}
\arguments{
\item{dataset}{The data to plot}
}
\description{
This gives a histogram overlaid with a normal l=plot and the kernel density
}
\examples{
sas.densityplot()
}

