% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.tplots.R
\name{sas.tplots}
\alias{sas.tplots}
\title{T Test Plots}
\usage{
sas.tplots(dataset, alpha = 0.05)
}
\arguments{
\item{dataset}{Data to plot}

\item{alpha}{Confidence level. Defaults to .05}
}
\description{
Combines density plot and boxplot to fit the SAS aesthetic
}
\examples{
sas.tplots()
}

