% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.ttables.R
\name{sas.ttables}
\alias{sas.ttables}
\title{T-Tables
Creates summary statistics for a vector and gives confidence interval for the mean and standard deviation}
\usage{
sas.ttables(dataset, alpha = 0.05)
}
\arguments{
\item{dataset}{The dataset to summarize}

\item{alpha}{The confidence level. Defaults to .05}
}
\description{
T-Tables
Creates summary statistics for a vector and gives confidence interval for the mean and standard deviation
}
\examples{
sas.ttables()
}

